refactor(dursto): move out default PersistentKeyValueStore impls
