+++ durable-storage/src/storage.rs
@@ -18,7 +18,6 @@ use tempfile::TempDir;
 use crate::commit::CommitId;
 use crate::errors::Error;
 use crate::errors::OperationalError;
-use crate::repo::DirectoryManager;
 
 /// Types that implement this trait can be used as the underlying key-value store
 pub trait KeyValueStore: Sized {
@@ -71,25 +70,8 @@ pub trait PersistentKeyValueStore: KeyValueStore + Sized {
     /// Implementations will treat the path as a directory.
     fn commit_to_path(&self, path: &Path) -> Result<(), OperationalError>;
 
-    /// Write the current key-value state to disk.
-    fn commit(&self, repo: &DirectoryManager, id: &CommitId) -> Result<(), OperationalError> {
-        let checkpoint_path = repo.database_commit_dir(id);
-
-        // If the path already exists, we overwrite the existing commit. This is highly unlikely to
-        // happen anyway if the commits are a hash of the content.
-        if Path::exists(&checkpoint_path) {
-            std::fs::remove_dir_all(&checkpoint_path).map_err(|error| {
-                OperationalError::DirRemovalFailed {
-                    path: checkpoint_path.clone(),
-                    error,
-                }
-            })?;
-
-            log::warn!("Overwriting existing commit: {}", id.hex_encode());
-        }
-
-        self.commit_to_path(&checkpoint_path)
-    }
+    /// Write the current key-value state to the repository.
+    fn commit(&self, repo: &Self::Repo, id: &CommitId) -> Result<(), OperationalError>;
 
     /// Checkout the state from `source_path` but leave it untouched. Modifications to the
     /// resulting state will be reflected in `working_path`.
@@ -98,13 +80,8 @@ pub trait PersistentKeyValueStore: KeyValueStore + Sized {
         working_path: TempDir,
     ) -> Result<Self, OperationalError>;
 
-    /// Retrieve a key-value state from disk.
-    fn checkout(repo: &DirectoryManager, id: &CommitId) -> Result<Self, OperationalError> {
-        let commit_path = repo.database_commit_dir(id);
-        let working_path = repo.temp_database_dir()?;
-
-        Self::checkout_from_path(&commit_path, working_path)
-    }
+    /// Retrieve a key-value state from the repository.
+    fn checkout(repo: &Self::Repo, id: &CommitId) -> Result<Self, OperationalError>;
 }
 
 #[cfg(test)]
